Compute Model Weights

Compute model weights from model ranks

model_weights_in_group(pred, groupvars = "model", rankvar = "rank")

Arguments

pred
data.frame that has at least columns specified via parameters groupvars and rank. pred is expected to have the same number of elements of grouping variable within each group. That is, each model should have been run for the same number of elements of groupvar. If this is not the case, use model_weights instead.

groupvars
quoted variable names used to define groups within which model weights are computed. Defaults to model.

rankvar
quoted variable names that contain model ranks.

Value

a data.frame with columns model and weight where weight is the unnormalised weight of the model

Details

In a group defined by groupvars, the weight of a model is defined as $$weight(M) = \sum\limits_{i = 1}^{M}{K(i) / i}$$ where \(K(i)\) is the number of times model M is ranked \(i\) among \(M_1, M_2, \dots M\) models.

See also

Examples

pred <- data.frame(
  model = as.factor(c("a", "b", "c", "a", "b", "c")),
  country = c("C1", "C1", "C1", "C2", "C2", "C2"),
  error = c(6L, 1L, 3L, 4L, 7L, 9L),
  date = c("1", "1", "1", "2", "2", "2"),
  stringsAsFactors = FALSE
)
ranked <- model_ranks(pred, c("country", "date"), "error")
model_weights_in_group(ranked)

#>   model weight npreds
#> a     a    1.3      2
#> b     b    1.5      2
#> c     c    0.8      2